Add PR fixes
