+++ common/src/main/org/jetbrains/lincheck/util/Utils.kt
@@ -28,7 +28,7 @@ val StackTraceElement.isLincheckInternals get() =
 
 internal const val LINCHECK_PACKAGE_NAME            = "org.jetbrains.lincheck."
 internal const val LINCHECK_KOTLINX_PACKAGE_NAME    = "org.jetbrains.kotlinx.lincheck."
-const val LINCHECK_RUNNER_PACKAGE_NAME     = "org.jetbrains.kotlinx.lincheck.runner."
+internal const val LINCHECK_RUNNER_PACKAGE_NAME     = "org.jetbrains.kotlinx.lincheck.runner."
 internal const val LINCHECK_BOOTSTRAP_PACKAGE_NAME  = "sun.nio.ch.lincheck."
 
 /**
