+++ integration-test/trace-debugger/src/main/org/jetbrains/trace/debugger/test/AbstractTraceDebuggerIntegrationTest.kt
@@ -8,12 +8,12 @@
  * with this file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
 
-package org.jetbrains.test.trace.recorder
+package org.jetbrains.trace.debugger.test
 
 import AbstractTraceIntegrationTest
 
-abstract class AbstractTraceRecorderIntegrationTest : AbstractTraceIntegrationTest() {
-    override val fatJarName: String = "trace-recorder-fat.jar"
+abstract class AbstractTraceDebuggerIntegrationTest : AbstractTraceIntegrationTest() {
+    override val fatJarName: String = "trace-debugger-fat.jar"
 
     final override fun runGradleTest(
         testClassName: String,
@@ -27,15 +27,12 @@ abstract class AbstractTraceRecorderIntegrationTest : AbstractTraceIntegrationTe
             testClassName,
             testMethodName,
             extraJvmArgs + listOf(
-                "-Dlincheck.traceRecorderMode=true",
+                "-Dlincheck.traceDebuggerMode=true",
                 "-XX:+UnlockExperimentalVMOptions",
                 "-XX:hashCode=2"
             ),
-            extraAgentArgs
-                .plus("text")
-                .plus("verbose"),
-            gradleCommands,
-            checkRepresentation
+            extraAgentArgs,
+            gradleCommands
         )
     }
 }
