+++ integration-test/trace-debugger/src/main/org/jetbrains/trace/debugger/test/KotlinImmutableCollectionsTraceDebuggerIntegrationTest.kt
@@ -8,7 +8,7 @@
  * with this file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
 
-package org.jetbrains.test.trace.debugger
+package org.jetbrains.trace.debugger.test
 
 import org.junit.Ignore
 import org.junit.Test
