+++ integration-test/trace-debugger/src/main/org/jetbrains/trace/debugger/test/TraceDebuggerExamplesTraceDebuggerIntegrationTest.kt
@@ -8,12 +8,13 @@
  * with this file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
 
-package org.jetbrains.test.trace.recorder
+package org.jetbrains.trace.debugger.test
 
+import org.junit.Ignore
 import org.junit.Test
 import java.nio.file.Paths
 
-class TraceDebuggerExamplesTraceRecorderIntegrationTest : AbstractTraceRecorderIntegrationTest() {
+class TraceDebuggerExamplesTraceDebuggerIntegrationTest: AbstractTraceDebuggerIntegrationTest() {
     override val projectPath: String = Paths.get("build", "integrationTestProjects", "TraceDebuggerExamples").toString()
 
     @Test
@@ -25,13 +26,13 @@ class TraceDebuggerExamplesTraceRecorderIntegrationTest : AbstractTraceRecorderI
         )
     }
 
-    @Test // actually runs the same test as above, but via the `runGradleTests` call
-    fun `org_examples_integration_bugs_LinkedHashSetTest`() {
-        runGradleTests(
-            testClassNamePrefix = "org.examples.integration.bugs.LinkedHashSetTest",
-            gradleBuildCommands = listOf(":compileTestKotlin"),
-            gradleTestCommands = listOf(":test"),
-            checkRepresentation = false,
+    @Ignore("`class.java.declaredMethods` call returns nondeterministic results")
+    @Test
+    fun `org_examples_integration_bugs_ReflectionTest sortMethods`() {
+        runGradleTest(
+            testClassName = "org.examples.integration.bugs.ReflectionTest",
+            testMethodName = "sortMethods",
+            gradleCommands = listOf(":test"),
         )
     }
 }
