+++ integration-test/trace-recorder/src/main/org/jetbrains/trace/recorder/test/AbstractTraceRecorderIntegrationTest.kt
@@ -8,12 +8,12 @@
  * with this file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
 
-package org.jetbrains.test.trace.debugger
+package org.jetbrains.trace.recorder.test
 
 import AbstractTraceIntegrationTest
 
-abstract class AbstractTraceDebuggerIntegrationTest : AbstractTraceIntegrationTest() {
-    override val fatJarName: String = "trace-debugger-fat.jar"
+abstract class AbstractTraceRecorderIntegrationTest : AbstractTraceIntegrationTest() {
+    override val fatJarName: String = "trace-recorder-fat.jar"
 
     final override fun runGradleTest(
         testClassName: String,
@@ -27,12 +27,15 @@ abstract class AbstractTraceDebuggerIntegrationTest : AbstractTraceIntegrationTe
             testClassName,
             testMethodName,
             extraJvmArgs + listOf(
-                "-Dlincheck.traceDebuggerMode=true",
+                "-Dlincheck.traceRecorderMode=true",
                 "-XX:+UnlockExperimentalVMOptions",
                 "-XX:hashCode=2"
             ),
-            extraAgentArgs,
-            gradleCommands
+            extraAgentArgs
+                .plus("text")
+                .plus("verbose"),
+            gradleCommands,
+            checkRepresentation
         )
     }
 }
