+++ integration-test/trace-recorder/src/main/org/jetbrains/trace/recorder/test/KotlinImmutableCollectionsTraceRecorderIntegrationTest.kt
@@ -8,7 +8,7 @@
  * with this file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
 
-package org.jetbrains.test.trace.recorder
+package org.jetbrains.trace.recorder.test
 
 import org.junit.Test
 import java.nio.file.Paths
